Restructure project
Groups classes by activities. The ui package is meant for common
fragments that may be shared across multiple activities
