+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/fragments/RollCallEventCreationFragment.java
@@ -1,4 +1,4 @@
-package com.github.dedis.student20_pop.ui.event.creation;
+package com.github.dedis.student20_pop.detail.fragments;
 
 import android.os.Bundle;
 import android.text.Editable;
