+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/fragments/pickers/package-info.java
@@ -3,4 +3,4 @@
  *
  * <p>The classes inside this package are fragments.
  */
-package com.github.dedis.student20_pop.ui.event.creation.pickers;
+package com.github.dedis.student20_pop.detail.fragments.pickers;
