+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/listeners/OnAddAttendeesListener.java
@@ -1,5 +1,6 @@
-package com.github.dedis.student20_pop.utility.ui.listener;
+package com.github.dedis.student20_pop.detail.listeners;
 
 public interface OnAddAttendeesListener {
+
   void onAddAttendeesListener(String eventId);
 }
