+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/detail/listeners/OnEventTypeSelectedListener.java
@@ -1,7 +1,8 @@
-package com.github.dedis.student20_pop.utility.ui.listener;
+package com.github.dedis.student20_pop.detail.listeners;
 
 import com.github.dedis.student20_pop.model.event.EventType;
 
 public interface OnEventTypeSelectedListener {
+
   void OnEventTypeSelectedListener(EventType eventType);
 }
