+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/HomeActivity.java
@@ -13,9 +13,9 @@ import com.github.dedis.student20_pop.Injection;
 import com.github.dedis.student20_pop.R;
 import com.github.dedis.student20_pop.ViewModelFactory;
 import com.github.dedis.student20_pop.detail.LaoDetailActivity;
-import com.github.dedis.student20_pop.ui.ConnectingFragment;
-import com.github.dedis.student20_pop.ui.HomeFragment;
-import com.github.dedis.student20_pop.ui.LaunchFragment;
+import com.github.dedis.student20_pop.home.fragments.ConnectingFragment;
+import com.github.dedis.student20_pop.home.fragments.HomeFragment;
+import com.github.dedis.student20_pop.home.fragments.LaunchFragment;
 import com.github.dedis.student20_pop.ui.qrcode.CameraPermissionFragment;
 import com.github.dedis.student20_pop.ui.qrcode.QRCodeScanningFragment;
 import com.github.dedis.student20_pop.utility.ActivityUtils;
