+++ fe2-android/app/src/main/java/com/github/dedis/student20_pop/home/listeners/LAOItemUserActionsListener.java
@@ -1,7 +1,8 @@
-package com.github.dedis.student20_pop.home;
+package com.github.dedis.student20_pop.home.listeners;
 
 import com.github.dedis.student20_pop.model.Lao;
 
 public interface LAOItemUserActionsListener {
+
   void onLAOClicked(Lao lao);
 }
