--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/utility/ui/adapter/AttendeeEventExpandableListViewAdapter.java
@@ -1,39 +1,0 @@
-package com.github.dedis.student20_pop.utility.ui.adapter;
-
-/** Adapter to show events of an Attendee */
-// public class AttendeeEventExpandableListViewAdapter extends EventExpandableListViewAdapter {
-
-//  /**
-//   * Constructor for the expandable list view adapter to display the events in the attendee UI
-//   *
-//   * @param context
-//   * @param events the list of events of the lao
-//   */
-//  public AttendeeEventExpandableListViewAdapter(Context context, List<Event> events) {
-//    super(context, events);
-//  }
-//
-//  /**
-//   * @param groupPosition
-//   * @param isExpanded
-//   * @param convertView
-//   * @param parent
-//   * @return the view for a given category
-//   */
-//  @Override
-//  public View getGroupView(
-//      int groupPosition, boolean isExpanded, View convertView, ViewGroup parent) {
-//    String eventCategory = ((EventCategory) getGroup(groupPosition)).toString();
-//
-//    if (convertView == null) {
-//      LayoutInflater inflater =
-//          (LayoutInflater) this.context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-//      convertView = inflater.inflate(R.layout.layout_event_category, null);
-//    }
-//
-//    TextView eventTextView = convertView.findViewById(R.id.event_category);
-//    eventTextView.setText(eventCategory);
-//
-//    return convertView;
-//  }
-// }
