--- fe2-android/app/src/main/java/com/github/dedis/student20_pop/utility/ui/listener/package-info.java
@@ -1,2 +1,0 @@
-/** Classes used to communicate between activities and its fragments */
-package com.github.dedis.student20_pop.utility.ui.listener;
