+++ fe2-android/app/src/main/res/layout/layout_lao_home.xml
@@ -1,54 +1,55 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android" xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <import type="android.view.View.OnClickListener" />
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto">
+  <data>
+    <import type="android.view.View.OnClickListener" />
 
-        <variable
-            name="lao"
-            type="com.github.dedis.student20_pop.model.Lao" />
+    <variable
+      name="lao"
+      type="com.github.dedis.student20_pop.model.Lao" />
 
-        <variable
-            name="listener"
-            type="com.github.dedis.student20_pop.home.LAOItemUserActionsListener" />
-    </data>
+    <variable
+      name="listener"
+      type="com.github.dedis.student20_pop.home.listeners.LAOItemUserActionsListener" />
+  </data>
 
-    <androidx.constraintlayout.widget.ConstraintLayout android:id="@+id/layout_lao_home"
-                                                       android:layout_width="match_parent"
-                                                       android:layout_height="wrap_content"
-                                                       android:background="@color/lao"
-                                                       android:onClick="@{() -> listener.onLAOClicked(lao)}"
-        >
+  <androidx.constraintlayout.widget.ConstraintLayout
+    android:id="@+id/layout_lao_home"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/lao"
+    android:onClick="@{() -> listener.onLAOClicked(lao)}">
 
-        <TextView
-                android:id="@+id/lao_name"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="@dimen/margin_text"
-                android:textSize="@dimen/size_body"
-                android:textStyle="bold"
-                android:text="@{lao.name}"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"/>
+    <TextView
+      android:id="@+id/lao_name"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_marginHorizontal="@dimen/margin_text"
+      android:textSize="@dimen/size_body"
+      android:textStyle="bold"
+      android:text="@{lao.name}"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-                android:id="@+id/date"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="@dimen/margin_text"
-                android:textSize="@dimen/size_small"
-                android:textStyle="italic"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"/>
+    <TextView
+      android:id="@+id/date"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_marginHorizontal="@dimen/margin_text"
+      android:textSize="@dimen/size_small"
+      android:textStyle="italic"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintTop_toTopOf="parent" />
 
-        <TextView
-                android:id="@+id/role"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="@dimen/margin_text"
-                android:textSize="@dimen/size_small"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/lao_name"/>
+    <TextView
+      android:id="@+id/role"
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_marginHorizontal="@dimen/margin_text"
+      android:textSize="@dimen/size_small"
+      app:layout_constraintStart_toStartOf="parent"
+      app:layout_constraintTop_toBottomOf="@+id/lao_name" />
 
-    </androidx.constraintlayout.widget.ConstraintLayout>
+  </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
